+++ atlasdb-config/src/test/java/com/palantir/atlasdb/config/AtlasDbConfigTest.java
@@ -47,27 +47,12 @@ public class AtlasDbConfigTest {
     private static final ServerListConfig SINGLETON_SERVER_LIST = ImmutableServerListConfig.builder()
             .addServers("server")
             .build();
-    private static final TimeLockClientConfig TIMELOCK_CONFIG = ImmutableTimeLockClientConfig.builder()
-            .client("client")
-            .serversList(SINGLETON_SERVER_LIST)
-            .build();
     private static final Optional<SslConfiguration> SSL_CONFIG = Optional.of(mock(SslConfiguration.class));
     private static final Optional<SslConfiguration> OTHER_SSL_CONFIG = Optional.of(mock(SslConfiguration.class));
     private static final Optional<SslConfiguration> NO_SSL_CONFIG = Optional.empty();
 
     private static final String TEST_NAMESPACE = "client";
     private static final String OTHER_CLIENT = "other-client";
-
-    private static final TimeLockClientConfig TIMELOCK_CONFIG_WITH_OPTIONAL_EMPTY_CLIENT = ImmutableTimeLockClientConfig
-            .builder()
-            .client(Optional.empty())
-            .serversList(SINGLETON_SERVER_LIST)
-            .build();
-    private static final TimeLockClientConfig TIMELOCK_CONFIG_WITH_OTHER_CLIENT = ImmutableTimeLockClientConfig
-            .builder()
-            .client(OTHER_CLIENT)
-            .serversList(SINGLETON_SERVER_LIST)
-            .build();
     private static final String CLIENT_NAMESPACE = "client";
 
     @BeforeClass
@@ -101,16 +86,6 @@ public class AtlasDbConfigTest {
     }
 
     @Test
-    public void configWithTimelockBlockIsValid() {
-        AtlasDbConfig config = ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG)
-                .build();
-        assertThat(config.getNamespaceString(), equalTo(TEST_NAMESPACE));
-        assertThat(config, not(nullValue()));
-    }
-
-    @Test
     public void remoteLockAndTimestampConfigIsValid() {
         AtlasDbConfig config = ImmutableAtlasDbConfig.builder()
                 .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
@@ -128,27 +103,6 @@ public class AtlasDbConfigTest {
                 .leader(LEADER_CONFIG)
                 .lock(SINGLETON_SERVER_LIST)
                 .timestamp(SINGLETON_SERVER_LIST)
-                .build();
-    }
-
-    @Test(expected = IllegalStateException.class)
-    public void timelockBlockNotPermittedWithLockAndTimestampBlocks() {
-        ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
-                .timelock(ImmutableTimeLockClientConfig.builder()
-                        .client("testClient")
-                        .serversList(SINGLETON_SERVER_LIST).build())
-                .lock(SINGLETON_SERVER_LIST)
-                .timestamp(SINGLETON_SERVER_LIST)
-                .build();
-    }
-
-    @Test(expected = IllegalStateException.class)
-    public void timelockBlockNotPermittedWithLeaderBlock() {
-        ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG)
-                .leader(LEADER_CONFIG)
                 .build();
     }
 
@@ -198,35 +152,6 @@ public class AtlasDbConfigTest {
     }
 
     @Test
-    public void absentNamespaceRequiresTimelockClient() {
-        assertThatThrownBy(() -> ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG_WITH_OPTIONAL_EMPTY_CLIENT)
-                .build())
-                .isInstanceOf(IllegalStateException.class)
-                .satisfies((exception) ->
-                        assertThat(exception.getMessage(), containsString("config should be present")));
-    }
-
-    @Test(expected = IllegalStateException.class)
-    public void absentNamespaceRequiresMatchingKvsNamespaceAndTimelockClient() {
-        ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG_WITH_OTHER_CLIENT)
-                .build();
-    }
-
-    @Test
-    public void namespaceAcceptsEmptyKvsNamespaceAndTimelockClient() {
-        AtlasDbConfig config = ImmutableAtlasDbConfig.builder()
-                .namespace("a client")
-                .keyValueService(KVS_CONFIG_WITHOUT_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG_WITH_OPTIONAL_EMPTY_CLIENT)
-                .build();
-        assertThat(config.getNamespaceString(), equalTo("a client"));
-    }
-
-    @Test
     public void inMemoryConfigCanHaveEmptyNamespace() {
         InMemoryAtlasDbConfig kvsConfig = new InMemoryAtlasDbConfig();
         assertFalse("This test assumes the InMemoryAtlasDbConfig has no namespace by default",
@@ -251,51 +176,10 @@ public class AtlasDbConfigTest {
     }
 
     @Test
-    public void inMemoryConfigCannotHaveEmptyNamespaceWithEmptyTimelockClient() {
-        InMemoryAtlasDbConfig kvsConfig = new InMemoryAtlasDbConfig();
-        assertFalse("This test assumes the InMemoryAtlasDbConfig has no namespace by default",
-                kvsConfig.namespace().isPresent());
-        assertThatThrownBy(() -> ImmutableAtlasDbConfig.builder()
-                .namespace(Optional.empty())
-                .keyValueService(kvsConfig)
-                .timelock(TIMELOCK_CONFIG_WITH_OPTIONAL_EMPTY_CLIENT)
-                .build())
-                .isInstanceOf(IllegalStateException.class)
-                .satisfies((exception) ->
-                        assertThat(exception.getMessage(), containsString("TimeLock client should not be empty")));
-    }
-
-    // Documenting that for in-memory, we don't care what the timelock client is - it just has to be non-empty.
-    @Test
-    public void inMemoryKeyspaceAndTimelockClientCanBeDifferent() {
-        InMemoryAtlasDbConfig kvsConfig = new InMemoryAtlasDbConfig();
-        assertFalse("This test assumes the InMemoryAtlasDbConfig has no namespace by default",
-                kvsConfig.namespace().isPresent());
-        ImmutableAtlasDbConfig config = ImmutableAtlasDbConfig.builder()
-                .keyValueService(kvsConfig)
-                .timelock(TIMELOCK_CONFIG_WITH_OTHER_CLIENT)
-                .build();
-        assertThat(config.getNamespaceString(), equalTo(OTHER_CLIENT));
-    }
-
-    @Test
-    public void namespaceAndTimelockClientShouldMatch() {
-        assertThatThrownBy(() -> ImmutableAtlasDbConfig.builder()
-                .namespace(TEST_NAMESPACE)
-                .keyValueService(KVS_CONFIG_WITHOUT_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG_WITH_OTHER_CLIENT)
-                .build())
-                .isInstanceOf(IllegalStateException.class)
-                .satisfies((exception) ->
-                        assertThat(exception.getMessage(), containsString("config should be the same")));
-    }
-
-    @Test
     public void namespaceAndKvsNamespaceShouldMatch() {
         assertThatThrownBy(() -> ImmutableAtlasDbConfig.builder()
                 .namespace(TEST_NAMESPACE)
                 .keyValueService(KVS_CONFIG_WITH_OTHER_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG_WITH_OPTIONAL_EMPTY_CLIENT)
                 .build())
                 .isInstanceOf(IllegalStateException.class)
                 .satisfies((exception) ->
@@ -323,16 +207,6 @@ public class AtlasDbConfigTest {
         AtlasDbConfig withSsl = AtlasDbConfigs.addFallbackSslConfigurationToAtlasDbConfig(withoutSsl, SSL_CONFIG);
         assertThat(withSsl.lock().get().sslConfiguration(), is(SSL_CONFIG));
         assertThat(withoutSsl.getNamespaceString(), equalTo(TEST_NAMESPACE));
-    }
-
-    @Test
-    public void addingFallbackSslAddsItToTimelockServersBlock() {
-        AtlasDbConfig withoutSsl = ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITH_NAMESPACE)
-                .timelock(TIMELOCK_CONFIG)
-                .build();
-        AtlasDbConfig withSsl = AtlasDbConfigs.addFallbackSslConfigurationToAtlasDbConfig(withoutSsl, SSL_CONFIG);
-        assertThat(withSsl.timelock().get().serversList().sslConfiguration(), is(SSL_CONFIG));
     }
 
     @Test
@@ -389,14 +263,4 @@ public class AtlasDbConfigTest {
                 .build()).isInstanceOf(IllegalStateException.class);
     }
 
-    @Test
-    public void canSpecifyZeroServersIfUsingTimelockBlock() {
-        ImmutableAtlasDbConfig.builder()
-                .keyValueService(KVS_CONFIG_WITHOUT_NAMESPACE)
-                .namespace(CLIENT_NAMESPACE)
-                .timelock(ImmutableTimeLockClientConfig.builder()
-                        .serversList(ImmutableServerListConfig.builder().build())
-                        .build())
-                .build();
-    }
 }