+++ atlasdb-config/src/test/java/com/palantir/atlasdb/factory/TransactionManagersTest.java
@@ -38,12 +38,10 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlMatching;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 import javax.ws.rs.core.Response;
 
@@ -62,7 +60,6 @@ import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedMap;
 import com.palantir.atlasdb.config.AtlasDbConfig;
@@ -70,11 +67,9 @@ import com.palantir.atlasdb.config.AtlasDbRuntimeConfig;
 import com.palantir.atlasdb.config.ImmutableAtlasDbConfig;
 import com.palantir.atlasdb.config.ImmutableLeaderConfig;
 import com.palantir.atlasdb.config.ImmutableServerListConfig;
-import com.palantir.atlasdb.config.ImmutableTimeLockClientConfig;
 import com.palantir.atlasdb.config.ImmutableTimeLockRuntimeConfig;
 import com.palantir.atlasdb.config.ImmutableTimestampClientConfig;
 import com.palantir.atlasdb.config.ServerListConfig;
-import com.palantir.atlasdb.config.TimeLockClientConfig;
 import com.palantir.atlasdb.factory.startup.TimeLockMigrator;
 import com.palantir.atlasdb.memory.InMemoryAtlasDbConfig;
 import com.palantir.atlasdb.qos.config.QosClientConfig;
@@ -147,7 +142,6 @@ public class TransactionManagersTest {
     private final MetricsManager metricsManager = MetricsManagers.createForTests();
 
     private int availablePort;
-    private TimeLockClientConfig mockClientConfig;
     private ServerListConfig rawRemoteServerConfig;
 
     private AtlasDbConfig config;
@@ -155,7 +149,6 @@ public class TransactionManagersTest {
     private Consumer<Object> environment;
     private TimestampStoreInvalidator invalidator;
     private Consumer<Runnable> originalAsyncMethod;
-    private Supplier<Optional<AtlasDbRuntimeConfig>> configSupplier;
 
     @ClassRule
     public static final TemporaryFolder temporaryFolder = new TemporaryFolder();
@@ -189,9 +182,9 @@ public class TransactionManagersTest {
 
         config = mock(AtlasDbConfig.class);
         when(config.leader()).thenReturn(Optional.empty());
+        when(config.namespace()).thenReturn(Optional.of(CLIENT));
         when(config.timestamp()).thenReturn(Optional.empty());
         when(config.lock()).thenReturn(Optional.empty());
-        when(config.timelock()).thenReturn(Optional.empty());
         when(config.keyValueService()).thenReturn(new InMemoryAtlasDbConfig());
         when(config.initializeAsync()).thenReturn(false);
 
@@ -206,11 +199,9 @@ public class TransactionManagersTest {
         when(invalidator.backupAndInvalidate()).thenReturn(EMBEDDED_BOUND);
 
         availablePort = availableServer.port();
-        mockClientConfig = getTimelockConfigForServers(ImmutableList.of(getUriForPort(availablePort)));
         rawRemoteServerConfig = ImmutableServerListConfig.builder()
                 .addServers(getUriForPort(availablePort))
                 .build();
-        configSupplier = () -> Optional.of(runtimeConfig);
     }
 
     @After
@@ -220,8 +211,7 @@ public class TransactionManagersTest {
 
     @Test
     public void userAgentsPresentOnRequestsToTimelockServer() {
-        setUpTimeLockBlockInInstallConfig();
-
+        setUpTimeLockBlockInRuntimeConfig();
         availableServer.stubFor(post(urlMatching("/")).willReturn(aResponse().withStatus(200).withBody("3")));
         availableServer.stubFor(TIMELOCK_LOCK_MAPPING.willReturn(aResponse().withStatus(200).withBody("4")));
 
@@ -313,7 +303,7 @@ public class TransactionManagersTest {
 
     @Test
     public void batchesRequestsIfBatchingEnabled() throws InterruptedException {
-        setUpTimeLockBlockInInstallConfig();
+        setUpTimeLockBlockInRuntimeConfig();
         when(runtimeConfig.timestampClient()).thenReturn(ImmutableTimestampClientConfig.of(true));
 
         createLockAndTimestampServicesForConfig(config, runtimeConfig).timestamp().getFreshTimestamp();
@@ -323,7 +313,7 @@ public class TransactionManagersTest {
 
     @Test
     public void doesNotBatchRequestsIfBatchingNotEnabled() {
-        setUpTimeLockBlockInInstallConfig();
+        setUpTimeLockBlockInRuntimeConfig();
         when(runtimeConfig.timestampClient()).thenReturn(ImmutableTimestampClientConfig.of(false));
 
         createLockAndTimestampServicesForConfig(config, runtimeConfig).timelock().getFreshTimestamp();
@@ -392,15 +382,14 @@ public class TransactionManagersTest {
 
     @Test
     public void metricsAreReportedExactlyOnceWhenUsingTimelockService() {
-        setUpTimeLockBlockInInstallConfig();
-
+        setUpTimeLockBlockInRuntimeConfig();
         assertThatTimeAndLockMetricsAreRecorded(TIMELOCK_SERVICE_FRESH_TIMESTAMP_METRIC,
                 TIMELOCK_SERVICE_CURRENT_TIME_METRIC);
     }
 
     @Test
     public void metricsAreReportedExactlyOnceWhenUsingTimelockServiceWithRequestBatching() {
-        setUpTimeLockBlockInInstallConfig();
+        setUpTimeLockBlockInRuntimeConfig();
         when(runtimeConfig.timestampClient()).thenReturn(ImmutableTimestampClientConfig.of(true));
 
         assertThatTimeAndLockMetricsAreRecorded(TIMESTAMP_SERVICE_FRESH_TIMESTAMP_METRIC,
@@ -431,12 +420,6 @@ public class TransactionManagersTest {
     }
 
     @Test
-    public void usesTimeLockIfInstallConfigIsUnspecifiedButInitialRuntimeConfigContainsTimeLockBlock() {
-        setUpTimeLockBlockInRuntimeConfig();
-        verifyUsingTimeLockByGettingAFreshTimestamp();
-    }
-
-    @Test
     public void throwsIfInstallConfigHasLeaderBlockButInitialRuntimeConfigContainsTimeLockBlock() throws IOException {
         setUpLeaderBlockInConfig();
         setUpTimeLockBlockInRuntimeConfig();
@@ -451,19 +434,9 @@ public class TransactionManagersTest {
     }
 
     @Test
-    public void usesTimeLockIfInstallConfigIsTimeLockAndInitialRuntimeConfigContainsTimeLockBlock() {
-        setUpTimeLockBlockInInstallConfig();
+    public void usesTimeLockIfInitialRuntimeConfigContainsTimeLockBlock() {
         setUpTimeLockBlockInRuntimeConfig();
         verifyUsingTimeLockByGettingAFreshTimestamp();
-    }
-
-    @Test
-    public void usesTimeLockIfInstallConfigIsTimeLockAndInitialRuntimeConfigDoesNotContainTimeLockBlock() {
-        setUpTimeLockBlockInInstallConfig();
-        verifyUsingTimeLockByGettingAFreshTimestamp();
-
-        assertTrue("Runtime config was not expected to contain a timelock block",
-                !runtimeConfig.timelockRuntime().isPresent());
     }
 
     @Test
@@ -525,10 +498,6 @@ public class TransactionManagersTest {
     private void setUpForRemoteServices() throws IOException {
         availableServer.stubFor(LEADER_UUID_MAPPING.willReturn(aResponse().withStatus(404)));
         setUpLeaderBlockInConfig();
-    }
-
-    private void setUpTimeLockBlockInInstallConfig() {
-        when(config.timelock()).thenReturn(Optional.of(mockClientConfig));
     }
 
     private void setUpTimeLockBlockInRuntimeConfig() {
@@ -607,15 +576,6 @@ public class TransactionManagersTest {
 
     private static String getUriForPort(int port) {
         return String.format("http://%s:%s", WireMockConfiguration.DEFAULT_BIND_ADDRESS, port);
-    }
-
-    private static TimeLockClientConfig getTimelockConfigForServers(List<String> servers) {
-        return ImmutableTimeLockClientConfig.builder()
-                .client(CLIENT)
-                .serversList(ImmutableServerListConfig.builder()
-                        .addAllServers(servers)
-                        .build())
-                .build();
     }
 
     private TransactionManagers.LockAndTimestampServices createLockAndTimestampServicesForConfig(